// src/app/models/project.ts
export class Project {
    "id": number;
    "projectName": string;
    "projectNr": string;
    "projectPath": string;
    "treeSortkey": string;
    "maxChildSortkey": string;
    "description": string;
    "billingTypeId": number;
    "note": string;
    "requiresReportP": boolean;
    "projectBudget": number;
    "projectRisk": string;
    "corporateSponsor": string;
    "percentCompleted": number;
    "projectBudgetHours": number;
    "costQuotesCache": number;
    "costInvoicesCache": number;
    "costTimesheetPlannedCache": number;
    "costPurchaseOrdersCache": number;
    "costBillsCache": number;
    "costTimesheetLoggedCache": number;
    "endDate": string;
    "startDate": string;
    "templateP": boolean;
    "sortOrder": number;
    "reportedHoursCache": number;
    "costExpensePlannedCache": number;
    "costExpenseLoggedCache": number;
    "confirmDate": string;
    "costDeliveryNotesCache": number;
    "costCacheDirty": string;
    "milestoneP": boolean;
    "releaseItemP": string;
    "presalesProbability": number;
    "presalesValue": number;
    "reportedDaysCache": number;
    "presalesValueCurrency": string;
    "opportunityPriorityId": number;
    "opportunitySalesStageId": number;
    "opportunityCampaignId": number;
    "scoreRevenue": number;
    "scoreStrategic": number;
    "scoreFinanceNpv": number;
    "scoreCustomers": number;
    "scoreFinanceCost": number;
    "costBillsPlanned": number;
    "costExpensesPlanned": number;
    "scoreRisk": number;
    "scoreCapabilities": number;
    "scoreEinanceRoi": number;
    "projectMembers": number;
    "projectUserwiseBoard": string;
    "projectBringNextday": number;
    "projectBringSameboard": string;
    "projectNewboardEverytime": string;
    "projectUserwiseBoard2": string;
    "projectBringSameboard2": string;
    "projectNewboard2Everytime": number;
    "projectNewboard2Always": string;
    "projectReportWeekly": string;
    "backlogPractice": number;
    "scoreGain": number;
    "scoreLoss": number;
    "scoreDelivery": number;
    "scoreOperations": number;
    "scoreWhy": number;
    "projectTheme": number;
    "projectClass": number;
    "projectVertical": number;
    "javaServices": string;
    "netServices": string;
    "collectionLink": string;
    "trainingLink": string;
    "collectionName": string;
    "trainingName": string;
    "trainingDoc": string;
    "testingRichtext": number;
    "templateCategory": string;
    "dType": number;
    "dOption": number;
    "dFilter": number;
    "dId": number;
    "tType": number;
    "tOption": number;
    "tFilter": number;
    "tId": number;
    "risktype": string;
    "riskimpact": number;
    "riskprobability": number;
    "projectInitiativeId": number;
    "projectBusinessgoalId": number;
    "projectSubgoalId": number;
    "projectMaingoalId": number;
    "projectBucketId": any;
    "projectCostCenterId": number;
    "projectBoardId": number;
    "projectBoard2Id": number;
    "projectStatusId": any;
    "projectTypeId": number;
    "projectLeadId": number;
    "parentId"?: any;
    "imEmployees": string;
}